pyproject.toml
[project]
name = "mswappinit"
version = "2025.04.22"
description = "a personal, opinionated python app initializer"
readme = "README.md"
authors = [{ name = "matt wartell", email = "matt.wartel@gmail.com" }]
requires-python = ">=3.12"
dependencies = ["loguru>=0.7.3", "pickledb>=1.3.2", "python-dotenv>=1.1.0"]

[project.urls]
"Homepage" = "https://github.com/mwartell/mswappinit"
"Source" = "https://github.com/mwartell/mswappinit"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = ["pytest>=8.3.5"]

[tool.ruff]
# Specify the directories or files to lint
src = ["src", "tests"]

# Enable specific rules or rule groups
lint.select = ["E", "F", "W", "C90"]

# Set the maximum line length
line-length = 88
